# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Claw Admin API
  description: Unified admin control plane for LLM/Skill/MCP config, SSO, organizations, users, and logs.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path
...
security:
  - BearerAuth: []
paths:
...
  /admin/llm/providers:
    get:
      tags: [admin-llm]
      summary: List LLM providers
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      responses:
        '200':
          description: List of providers (API key masked)
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin-llm]
      summary: Create provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderCreate' }
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/llm/providers/{id}:
    get:
      tags: [admin-llm]
      summary: Get provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Provider (key masked)
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [admin-llm]
      summary: Update provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderUpdate' }
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin-llm]
      summary: Delete provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/sso/providers:
    get:
      tags: [admin-sso]
      summary: List SSO providers
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      responses:
        '200':
          description: List (clientSecret masked)
    post:
      tags: [admin-sso]
      summary: Create SSO provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [provider, issuer, clientId]
              properties:
                provider: { type: string, enum: [oidc, saml] }
                name: { type: string }
                issuer: { type: string }
                clientId: { type: string }
                clientSecret: { type: string }
                config: { type: object }
      responses:
        '201':
          description: Created
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/sso/providers/{id}:
    get:
      tags: [admin-sso]
      summary: Get SSO provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Provider (secret masked)
    put:
      tags: [admin-sso]
      summary: Update SSO provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                issuer: { type: string }
                clientId: { type: string }
                clientSecret: { type: string }
                config: { type: object }
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [admin-sso]
      summary: Delete SSO provider
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'

  # ---------- Client (read/report) ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      description: Tenant slug (optional; can come from JWT org_id)
      schema:
        type: string
        example: default
  schemas:
...
    ProviderCreate:
      type: object
      required: [providerId]
      properties:
        orgId: { type: string, nullable: true }
        providerId: { type: string }
        displayName: { type: string }
        baseUrl: { type: string }
        apiKey: { type: string }
        extraModels: { type: array, items: { type: object, properties: { id: { type: string }, name: { type: string } } } }
        isActive: { type: boolean }
        activeModel: { type: string }
    ProviderUpdate:
      type: object
      properties:
        displayName: { type: string }
        baseUrl: { type: string }
        apiKey: { type: string }
        extraModels: { type: array }
        isActive: { type: boolean }
        activeModel: { type: string }
  responses:
...
    Forbidden:
      description: Forbidden (admin role required)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: FORBIDDEN }
                  message: { type: string }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: VALIDATION_ERROR }
                  message: { type: string }
                  details: { type: object }
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: NOT_FOUND }
                  message: { type: string }